refactor(architecture): establish core project modules
